structural changes
structural changes in entry point of application
