+++ webapp/src/index.css
@@ -1,17 +1,1 @@
-html,
-body,
-#root {
-    height: 100%;
-    width: 100%;
-}
 
-body {
-    font-family: system-ui;
-    margin: 0;
-}
-
-*,
-*:after,
-*:before {
-    box-sizing: border-box;
-}
